requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv>=1.0.0
langgraph>=0.2.0
langgraph-checkpoint-mongodb>=0.3.0
langgraph-store-mongodb>=0.1.0
langchain>=0.3.0
langchain-mongodb>=0.2.0
pymongo>=4.6.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
sentence-transformers>=2.2.0
numpy>=1.24.0
